Battleship Game
Battleship Game. Creates a 5x5 Board, fill it with random number of ship (between 0 - 3) in different rows and with different sizes (one per size). Paint it. The player has 10 turns for winning, program controls the number of boats and their sizes and if they are sunken or not during the game.