Django doesnt allow to use Shared Primary Keys. Patch for solving it.
We want to show all records coming from a table in a template.Plus, we want to modify their values directly over the template. When table has Shared Primary Key Django doesnt allow to use it in his model. For solving it we create our own template and view and call a external query.
Be warn about how we solve the problem when you want to update multiples rows, you´ll need the old values, stored in the checkbox, and the new ones which be recover from the textbox.
You can use this idea in those tables which doesnt fix with Django model. You can improve theses querys using less through join, where, in and lists.