Own Paginator
Paginator shows, if it is possible, 15 pages, with the actual one centrates in those 15 pages. Always showing the first and the last ones of the total. Showing a message with actual page, number total of pages, actual records, number total of records.
Aplying different styles for actual page, another ones, first one and last one, and page in hover.